use class methods instead of overriding relation
