--- lib/spyke/kaminari/relation.rb
@@ -1,54 +1,0 @@
-module Spyke::Kaminari
-  # A Spyke relation extended with pagination metadata and helpers, similar
-  # Kaminari's PaginatableArray.
-  class Relation < Spyke::Relation
-    Spyke::Kaminari::HEADERS.keys.each do |symbol|
-      define_method(symbol) do
-        find_some.metadata[Spyke::Kaminari::METADATA_KEY][symbol]
-      end
-    end
-
-    def first_page?
-      current_page == 1
-    end
-
-    def last_page?
-      current_page == total_pages
-    end
-
-    def out_of_range?
-      current_page > total_pages
-    end
-
-    # Iterates through each page.
-    #
-    # Useful for stitching together an array of all records, across every page.
-    # For example:
-    #
-    #   users = User.all.each_page.flat_map(&:to_a)
-    #
-    # If there are 3 total pages, it would result in these HTTP requests:
-    #
-    #   GET /v1/users
-    #   GET /v1/users?page=2
-    #   GET /v1/users?page=3
-    #
-    # ...and this array of users (25 records per page):
-    #
-    #   users.count
-    #   # => 75
-    #
-    # With the exception of the first one, pages are lazily fetched.
-    def each_page
-      return to_enum(:each_page) unless block_given?
-
-      relation = clone
-      yield relation
-      return if relation.last_page? || relation.out_of_range?
-
-      (relation.next_page..relation.total_pages).each do |number|
-        yield clone.page(number)
-      end
-    end
-  end
-end
